Add focus rows
